% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{computeLatitudePrecision}
\alias{computeLatitudePrecision}
\title{Longitude normalization}
\usage{
computeLatitudePrecision(codeLength)
}
\arguments{
\item{codeLength}{A number.}
}
\value{
A number.
}
\description{
Calculates the latitude precision value for a given code length.
Lengths <= 10 have the same precision for latitude and longitude,
but lengths > 10 have different precisions due to the grid method having fewer columns than rows.
}
\examples{
computeLatitudePrecision(10)
computeLatitudePrecision(12)
}
