% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{normalizeLatitude}
\alias{normalizeLatitude}
\title{Latitude normalization}
\usage{
normalizeLatitude(latitude)
}
\arguments{
\item{latitude}{A number.}
}
\value{
A number.
}
\description{
Limit latitude within coherence range (-90, 90)
}
\examples{
normalizeLatitude(80)
normalizeLatitude(100)
}
