% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{normalizeLongitude}
\alias{normalizeLongitude}
\title{Longitude normalization}
\usage{
normalizeLongitude(longitude)
}
\arguments{
\item{longitude}{A number.}
}
\value{
A number.
}
\description{
Limit longitude within coherence range (-180, 180)
}
\examples{
normalizeLongitude(100)
normalizeLongitude(181)
}
